Create BusinessCaseModel
In a community of People and a Bank, people own bank accounts for two reasons, either to just save their money or to invest their money or for both reasons on a separate record-line (Savings Account and Investment Account). Under saving accounts, cash can be deposited and withdrawn at anytime during the bank’s working hours. Withdrawals via the ATM machine operates on a 24/7 basis for the savings account. Under Investment accounts, the investor agrees with the bank on the date of refund and withdrawals can only be done on or after that date. Monthly interest rate will be determined by the bank depending on the amount of money invested and duration of investment. All activities of these accounts should be operational and accessible via a modulated mobile app that connects to MTN and Orange Mobile money accounts which permits both deposits and withdrawals from the bank.
The Bank should have a manager and customer service modules to manage and monitor the operations of as above.